+++ src/index.ts
@@ -2,7 +2,9 @@
 
 export * from './api.schema.js';
 export * from './base-params.js';
+export * from './progress.schema.js';
 export * from './recipe-template.schema.js';
 export * from './recipe.schema.js';
 export * from './request.schema.js';
 export * from './response.schema.js';
+export * from './webhook.schema.js';