+++ src/response.schema.ts
@@ -2,7 +2,7 @@
 
 import * as z from "zod/v4";
 
-export const PublisherResponse = z.object({
+export const PublishResponse = z.object({
 	job_id: z.string().regex(/^\d+$/)
 		.describe('Unique identifier for this job, can be used to query the status of the job.'),
 	reference_id: z.string().max(255)
@@ -11,4 +11,4 @@ export const PublisherResponse = z.object({
 		.describe('ISO datetime of the job.'),
 })
 	.describe('Publisher job output');
-export type PublisherResponse = z.infer<typeof PublisherResponse>;
+export type PublishResponse = z.infer<typeof PublishResponse>;